backend/requirements.txt
fastapi
python-dotenv
chromadb
sentence-transformers
uvicorn
